eMapMode: Move from enums.h into maps.h
